# Conflicts:
#	ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/application/transactional/ApplicationNotificationServiceImpl.java
#	ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/invite/transactional/ApplicationInviteServiceImpl.java
#	ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/project/monitoringofficer/transactional/MonitoringOfficerServiceImpl.java
#	ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/user/transactional/RegistrationServiceImpl.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/application/transactional/ApplicationNotificationServiceImplTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/assessment/transactional/AssessmentInviteServiceImplTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/assessment/transactional/AssessorServiceImplTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/competitionsetup/transactional/CompetitionSetupServiceImplTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/interview/transactional/InterviewAllocationServiceImplTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/invite/transactional/ApplicationInviteServiceImplTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/invite/transactional/InviteUserServiceImplTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/project/grantofferletter/transactional/GrantOfferLetterServiceImplTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/review/transactional/ReviewInviteServiceImplTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/thread/service/FinanceCheckQueriesServiceTest.java
#	ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/user/transactional/RegistrationServiceImplTest.java
#	ifs-web-service/ifs-application-service/src/test/java/org/innovateuk/ifs/registration/RegistrationServiceImplTest.java

+++ ifs-web-service/ifs-web-core/src/main/java/org/innovateuk/ifs/application/finance/view/FinanceViewHandlerProvider.java
@@ -27,7 +27,7 @@ public class FinanceViewHandlerProvider implements FinanceHandlerProvider {
     private JESFinanceFormHandler jesFinanceFormHandler;
 
     public FinanceFormHandler getFinanceFormHandler(long organisationType) {
-        if(RESEARCH.getId().longValue() == organisationType) {
+        if(RESEARCH.getId() == organisationType) {
             return jesFinanceFormHandler;
         } else {
             return defaultFinanceFormHandler;
@@ -35,7 +35,7 @@ public class FinanceViewHandlerProvider implements FinanceHandlerProvider {
     }
 
     public FinanceFormHandler getProjectFinanceFormHandler(long organisationType) {
-        if(RESEARCH.getId().longValue() == organisationType) {
+        if(RESEARCH.getId() == organisationType) {
             return jesFinanceFormHandler;
         } else {
             return projectFinanceFormHandler;
@@ -43,7 +43,7 @@ public class FinanceViewHandlerProvider implements FinanceHandlerProvider {
     }
 
     public FinanceModelManager getFinanceModelManager(long organisationType) {
-        if(RESEARCH.getId().longValue() == organisationType) {
+        if(RESEARCH.getId() == organisationType) {
             return getJESFinanceModelManager();
         } else {
             return getDefaultFinanceModelManager();
@@ -51,7 +51,7 @@ public class FinanceViewHandlerProvider implements FinanceHandlerProvider {
     }
 
     public FinanceModelManager getProjectFinanceModelManager(long organisationType) {
-        if(RESEARCH.getId().longValue() == organisationType) {
+        if(RESEARCH.getId() == organisationType) {
             return getJESProjectFinanceModelManager();
         } else {
             return getDefaultProjectFinanceModelManager();
